Added subscript, call & attrib access to the parser
With these additions, the parser is now capable of parsing the full example.bdf draft document.

Also modified the draft to remove the 'magic @ offset' syntax in favor of a mapping. Syntactic sugar for this
is considered unnecessary at this point.

Next steps are fixing up the parser by handling edge cases, adding documentation & tests for bindef, and making
the features added in this commit actually function.