feat(bindef): add field validations and conditions
Introduce field validation via the "valid" key of a format type which contains
a lazily-evaluated expression that determines whether the field is valid.

Introduce conditional fields via the "if" key which also contains a lazily-evaluated
expression determining whether the field is to be parsed.

An example of these new features can be found in the elf.bdf file and in the updated
reference spec.